Recommitted stuff...becuase of corruption.
I had about ten commits that couldn't be pushed becuase of a corruption
that happened somehwere. It was going to be really hard to fix it. So, I
just deleted everything and copied the files back in. This cuases a loss
of the commits log, but the work is preserved. I can live with that for
now. The commits included optimizer and vs template improvements as well
as some minor fixes in durandal itself.