Re-enabled parameters on auto-routes.
I think this can be useful. I implemented it slightly better this time I
think. The params object passed to the activate method on navigation
will now have a splat property only if additional route paramters were
found. The splat will be an array of those parameters in order that they
appeared. If there are not params, splat will be undefined. This makes
it easy to check and differentiates it from names parameters from an
explicitly mapped route.