Enabled showing modals via module id.
app.showModal and modalDialog.show now accept either a module (as
before) or a moduleId (new0 as their first parameter. If the parameters
is a string (moduleId) then we will automatically resolve the module. If
the module is a function, we will call it with new. Otherwise, we assume
that it is the modal model itself.