Fixing css issues
Fixed issue with the sequence of loading the css files. If you load the
app.css before the bootstrap responsive it creates a weird gap at the
top of the page when you size down. Also fixed issue with the bundle not
have the correct root directory which prevents the font awesome fonts
from being downloaded when the bundling is on.