Composition no longer animates to same view.
If the composition engine detects that the same view is being composed
as the active view, it will not perform the transition animation.
Furthermore, there is a new composition setting called
skipTransitionOnSameViewId which also allows the developer to indicated
that even if the views are different instances...if their ids are the
same (meaning they have the same html definition file) then also skip
the animation.