Added a new router binding handler.
The new router binding handler requires the installation of the router
as a plugin. The new handler simplifies the router binding process by
automatically hooking up the router by first looking for a router
property on the binding, second looking up a router property on the view
model and finally falling back to the root router. See the samples for
example usage.