The plugins api has changed to support multiple locations.
Instead of setting a plugins property on the app, you call
configurePlugins. The first parameter is the same as before, an object
with a property per plugin, optionally specifying its confguration. The
configurePlugins API has an optional second parameter which you can use
to specify the baseUrl to load the plugins from. By default it is
'plugins/'