Added router.navigatingBack to the main router.
As it turns out, there's no way to tell whether or not the user
navigated back or forward. This property is a best guess. It will be
true if the interaction did not happen through a link or navigate api
call and if the previous and new route fragments do not match. Note, we
cannot tell the difference between forward/back button presses. Also, we
cannot tell if someone typed a new url in their browser as opposed to
using the forward/back buttons. In both of these cases, the router will
have navigatingBack true.