Changed the .net projects back to using standard script inclusion.
After some work with various clients and others in the industry I
decided that for the .net starter kit and samples it was best to keep it
closer in configuration to 1.x. So, jQuery, Knockout and Bootstrap are
included via standard script tags rather than using require.js paths and
forcing shimming. Durandal itself still assumes no global libraries so
two define calls to push jQuery and Knockout into the AMD system are
required.