BREAKING CHANGE to system.acquire
Previously when passing multiple module ids to system aqcuire, the
promise would return resolve with one argument per module. However, this
causes problems when used with non-jQuery promises. As a result the
acquire function was reworked so that if an array of ids is passed or
multiple ids are passed then the promise resolves with an array of
modules.