Minor break to router.navigate and history.navigate.
I received a lot of feedback that requireing to pass true as the second
parameter to navigate in order to trigger the route handler was not
intuitive. As a result, tiggering will now be the default. To not
trigger, pass false as the second parameter. Otherwise the options
object works the same as before.