Remove *details part when hash generated
Default implementation of convertRouteToHash doesn't remove "*details" (or such patterns when working with child routers) from the end of routes. Such "*details" parts are not usable in any scenario, but removing them makes developer's work easy (Specially when we have child routers).

I know i can change this default implementation form outside, but it's messy to edit that for both root-router and child routers (Let me know if i'm wrong, but replacing these hooks on entire module only replaces them in root-router and i'll have to replace it for each child router too).

--
As a side note, it's almost a year that I haven't been in touch with this framework and SPA development, and now that i'm back I see it has become amazing. Thanks a lot for all the hard work.