Add optional change detection
When using the observable plugin, add an option to register a method on
each object to automatically detect changes. This method receives as
parameters: the obj that has changed, the name of the changed property,
and either the newValue if it's a value type, or the arrayChange object
if it's an array.