Added custom error handler in composition engine
It started small with only require.js load error (which is what I
needed), but I extended it to the whole module. Biggest change was
passing around the context and element to the whole stack so it's always
available.

Very useful to managing load error elegantly instead of crashing the
application.