Removed VSIX support.
It wasn't really much different from the nuget starter kit, so there
isn't much value in having it. It was extremely...let me say that
again...extremely painful to maintain. It was the most dreaded part of
the release process and I don't exaggerate when I say that it has
actually had an effect on release cadence for this project. VSIX
creation of this sort is not well documented. The process is error
prone. I don't want to have anything to do with it.