To hide and show, set css visibility so offsetHeight and offsetWidth work
When knockout bindings are executing during composition the view is currently display:none
This causes an issue if you need to calculate the width of elements on the page.  

In my case I was trying to use blockUI to mask an area of the view which loads after the view is attached.  BlockUI uses the element's width/height to center itself in the element.  

The proposed change merely changes the composition engine to set visibility instead of display none, and I also removed the calls to set knockout domdata since I searched the codebase and this was the only use of it and it doesn't make sense to use when the css property has only 2 possibilities.