Fixing compostion callback issue
Composition callbacks are called even if the last composition had an error. It'll give a consistent behavior as all the composition callbacks will be called properly. Though if there's an error in one of the composition callback then not all the composition callbacks will be called, for example router plugin can be in an inconsistent state after that.