Grand child routers parents isNavigation is false
When having 3 or more routers involved in a router, then isNavigation for the root will change to false after the 1st child routers view has composition complete. It should not register completed until we have seen if there is a child router that also needs to process additional routing information.