Fix in history's checkUrl
There are some edge cases (that we've experienced) in which
`history.fragment` should be equal to `getFragment()` result, but it
doesn't match because of some characters that aren't properly encoded.
This happens even if `getFragment()` used `history.getHash()`. This
solution solves it for our use cases, and the characters being replaced
are based on those reported in this website:

<http://www.justarrangingbits.org/firefox-magic-decoding-address-bar/index.html>